Additional configuration, load /.parity.yml
Adds two new configuration keys:

  Configuration#heroku_app_name
    defaults to "%{basename}-%{environment}"

  Configuration#heroku_app_basename
    defaults to "%{environment}"

Default operation w.r.t. this and existing attributes should be fully
backwards compatible.

If ./.parity.yml is present, load it. For each key found, set the
corresponding value in Configuration. This combined with the above
allows for easy customization of the app and remote names even when they
don't follow the `-environment` convention.